refactor: 상수들을 utils 폴더에서 최상위 constants 디렉토리로 이동
